♻️ Refactor : Apply changed path
