+++ src/main/java/com/run_us/server/domains/user/domain/AuthResult.java
@@ -1,8 +1,5 @@
 package com.run_us.server.domains.user.domain;
 
-import lombok.Getter;
-
-@Getter
 public record AuthResult (AuthResultType type, TokenPair tokenPair) {
 
 }
