+++ src/main/java/com/run_us/server/domains/user/domain/TokenPair.java
@@ -1,8 +1,5 @@
 package com.run_us.server.domains.user.domain;
 
-import lombok.Getter;
-
-@Getter
 public record TokenPair(String accessToken, String refreshToken) {
 
 }
